# bin/Debug/net8.0/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "MisticFy",
    "version": "v1"
  },
  "paths": {
    "/Login/authorize": {
      "get": {
        "tags": [
          "Login"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Login/callback": {
      "get": {
        "tags": [
          "Login"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Playlist/userPlaylist": {
      "get": {
        "tags": [
          "Playlist"
        ],
        "parameters": [
          {
            "name": "userPlaylist",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifyPlaylistDetailsDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifyPlaylistDetailsDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifyPlaylistDetailsDTO"
                }
              }
            }
          }
        }
      }
    },
    "/Playlist/AddSongToPlaylist": {
      "post": {
        "tags": [
          "Playlist"
        ],
        "parameters": [
          {
            "name": "playlistId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTrackRequestDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTrackRequestDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddTrackRequestDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifyPlaylistDetailsDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifyPlaylistDetailsDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifyPlaylistDetailsDTO"
                }
              }
            }
          }
        }
      }
    },
    "/Playlist/CreatePlaylist": {
      "post": {
        "tags": [
          "Playlist"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Playlist"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Playlist"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Playlist"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifyPlaylistDetailsDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifyPlaylistDetailsDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifyPlaylistDetailsDTO"
                }
              }
            }
          }
        }
      }
    },
    "/Search/search-track": {
      "get": {
        "tags": [
          "Search"
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifySearchResultTrackDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifySearchResultTrackDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifySearchResultTrackDTO"
                }
              }
            }
          }
        }
      }
    },
    "/Search/search-playlist": {
      "get": {
        "tags": [
          "Search"
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifySearchResultPlaylistDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifySearchResultPlaylistDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifySearchResultPlaylistDTO"
                }
              }
            }
          }
        }
      }
    },
    "/Search/search-artist": {
      "get": {
        "tags": [
          "Search"
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifySearchResultArtistDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifySearchResultArtistDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifySearchResultArtistDTO"
                }
              }
            }
          }
        }
      }
    },
    "/Search/search-album": {
      "get": {
        "tags": [
          "Search"
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifySearchResultAlbumDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifySearchResultAlbumDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifySearchResultAlbumDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddTrackRequestDTO": {
        "type": "object",
        "properties": {
          "tracks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpotifyMusicDTO"
            },
            "nullable": true
          },
          "positionToInsert": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Followers": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string",
            "nullable": true
          },
          "total": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "Image": {
        "type": "object",
        "properties": {
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "width": {
            "type": "integer",
            "format": "int32"
          },
          "url": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Playlist": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isPublic": {
            "type": "boolean"
          },
          "musics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpotifyMusicDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PublicUser": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "externalUrls": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          },
          "followers": {
            "$ref": "#/components/schemas/Followers"
          },
          "href": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Image"
            },
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "uri": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SpotifyAlbumDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "releaseDate": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SpotifyArtistDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SpotifyMusicDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "artists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpotifyArtistDTO"
            },
            "nullable": true
          },
          "album": {
            "$ref": "#/components/schemas/SpotifyAlbumDTO"
          },
          "uri": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SpotifyPlaylistDetailsDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "owner": {
            "$ref": "#/components/schemas/PublicUser"
          },
          "musics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpotifyMusicDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SpotifySearchResultAlbumDTO": {
        "type": "object",
        "properties": {
          "albums": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpotifyAlbumDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SpotifySearchResultArtistDTO": {
        "type": "object",
        "properties": {
          "artists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpotifyArtistDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SpotifySearchResultPlaylistDTO": {
        "type": "object",
        "properties": {
          "playlists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpotifyPlaylistDetailsDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SpotifySearchResultTrackDTO": {
        "type": "object",
        "properties": {
          "tracks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpotifyMusicDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Insira o token JWT no formato: Bearer {seu_token}",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}