# src/DTO/SearchMap/SpotifySearchResultAlbumDTO.cs
using MisticFy.API.src.DTO.SpotifyDTO;

namespace MisticFy.API.src.DTO.SearchMap
{
    public class SpotifySearchResultAlbumDTO
    {
        public List<SpotifyAlbumDTO> Albums { get; set; } = new();
    }
}