# src/DTO/SearchMap/SpotifySearchResultArtistDTO.cs
using MisticFy.API.src.DTO.SpotifyDTO;

namespace MisticFy.API.src.DTO.SearchMap
{
    public class SpotifySearchResultArtistDTO
    {
        public List<SpotifyArtistDTO> Artists { get; set; } = new();
    }
}